[Refactor] 대시보드, 마이페이지 패키지 구조 변경
## 연관된 이슈
#24
