+++ src/main/java/com/fcc/PureSync/repository/MenuRepository.java
@@ -1,7 +1,7 @@
 package com.fcc.PureSync.repository;
 
 import com.fcc.PureSync.entity.Menu;
-import com.fcc.PureSync.vo.MenuStatsNativeVo;
+import com.fcc.PureSync.context.dashboard.vo.MenuStatsNativeVo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
