--- src/main/java/com/fcc/PureSync/vo/EmotionNativeVo.java
@@ -1,6 +1,0 @@
-package com.fcc.PureSync.vo;
-
-public interface EmotionNativeVo {
-    String getDate();
-    int getEmoCode();
-}
